+++ plugins/12_Midnight/localization/deDE.lua
@@ -33,6 +33,10 @@ L['options_icons_profession_treasures_desc'] = 'Zeigt die Positionen der Schätz
 L['options_icons_safari'] = '{achievement:61091}'
 L['options_icons_safari_desc'] = 'Zeigt die Positionen der Kampfhaustiere für den Erfolg {achievement:61091} an.'
 
+L['options_icons_renowned_beast'] = 'Renowned Beasts'
+L['options_icons_renowned_beast_desc'] = 'Display locations of summonable daily beasts for Skinning'
+L['silverscale_note'] = 'Summoned below the bridge.'
+
 -------------------------------------------------------------------------------
 -------------------------------- EVERSONG WOODS --------------------------------
 -------------------------------------------------------------------------------
@@ -109,6 +113,22 @@ L['options_icons_put_a_pin_in_it'] = '{achievement:62199}'
 L['options_icons_put_a_pin_in_it_desc'] = nil
 L['kalika_note'] = nil
 L['songseeker_ikaja_note'] = nil
+
+-------------------------------------------------------------------------------
+----------------------------------- NAIGTAL ------------------------------------
+-------------------------------------------------------------------------------
+
+L['sleepy_mandrake_note'] = nil
+L['ancient_crypt_reliquary'] = nil
+L['squirming_mollusk'] = nil
+L['spiked_shell'] = nil
+
+-------------------------------------------------------------------------------
+------------------------------------- VAL --------------------------------------
+-------------------------------------------------------------------------------
+
+L['enchanted_hilt'] = nil
+L['enchanted_hilt_note'] = nil
 
 -------------------------------------------------------------------------------
 ---------------------------------- ARCANTINA ----------------------------------